Update pokedex.csv
Adding in all the alolan forms. The best way I could think of to make them distinct from their Kanto forms was to make them in the 1000s because there are only 800 pokemon now. If you have a better way to do it, go ahead, but I thought you would probably like the numbers to be integers.