Adding new tier 6 raid & realism factor for trainers2win
Mewtwo as a normal raid boss has 150% the hp of normal legendary raids, so I am adding a 6th raid tier. Also, instead of accounting for pokemon fainting, lag and other errors in raid battles by subtracting a number of seconds from the raid time when calculating the number of trainers necessary to win, I added a factor to decrease the actual DPS output of the pokemon in the calculation.